Musime spravit write single register cez vlastnu implmentaciu pretoze ocividne v plsim nam to nepojde ... takze teraz nam treba dorobit read single register aby sme mohli dokomcit write

+++ src/Server/ModbusClient.h
@@ -71,7 +71,7 @@ private:
 public:
     ModbusClient();
     ~ModbusClient() = default;
-    uint8_t ReadSingleCoil();
+    uint16_t ReadSingleCoil();
     bool WriteSingleCoil(const bool &Data,const int Byte,const int Bit);
 };
